+++ src/Dependencies.lua
@@ -6,15 +6,21 @@ require('src.libs.StateMachine')
 require('src.libs.StateStack')
 _G.suit = require('src.libs.SUIT')
 
-require('src.constants')
-
 gFonts = {
-    ['pixel'] = love.graphics.newFont('assets/FortAvenue-nAWrg.ttf', 16),
+    ['large'] = love.graphics.newFont('assets/FortAvenue-nAWrg.ttf', 32),
+    ['medium'] = love.graphics.newFont('assets/FortAvenue-nAWrg.ttf', 16),
+    ['small'] = love.graphics.newFont('assets/FortAvenue-nAWrg.ttf', 8),
 }
 
 gFrames = {
-    ['CoffeeMachine'] = love.graphics.newImage('assets/coffeeMachine.jpg')
+    ['CoffeeMachine'] = love.graphics.newImage('assets/coffeeMachine.jpg'),
+    customers = {
+        love.graphics.newImage('assets/Customer1.png'), --Grumpy Old Man
+    }
 }
+
+require('src.constants')
+
 
 require('src.states.BaseState')
 require('src.states.entity.CustomerState')
@@ -23,4 +29,5 @@ require('src.states.game.StartMenu')
 require('src.states.game.DayEndState')
 
 require('src.states.entity.CoffeeMachine')
-require('src.states.entity.Customer')
+require('src.states.entity.CustomerState')
+require('src.states.entity.Cursor')