+++ src/constants.lua
@@ -4,6 +4,30 @@ VIRTUAL_WIDTH = 432
 VIRTUAL_HEIGHT = 243
 
 COFFEE_MACHINE_ENTITY = {
+    frame = gFrames['CoffeeMachine'],
     x = 10,
-    y = 25,
+    y = 130,
+    desired_width = 32,
+    desired_height = 32,
+}
+
+CUSTOMER_ENTITIES = {
+    ['left'] = {
+        x = 10,
+        y = 25,
+        desired_width = 64,
+        desired_height = 64,
+    },
+    ['center'] = {
+        x = 110,
+        y = 25,
+        desired_width = 64,
+        desired_height = 64,
+    },
+    ['right'] = {
+        x = 210,
+        y = 25,
+        desired_width = 64,
+        desired_height = 64,
+    },
 }