+++ src/states/entity/CoffeeMachine.lua
@@ -1,7 +1,9 @@
 CoffeeMachine = class {__includes = BaseState}
 
-function CoffeeMachine:enterParams(params)
-    self = params
+function CoffeeMachine:init(params)
+    for k, v in pairs(params) do
+        self[k] = v
+    end
 end
 
 function CoffeeMachine:update(dt)
@@ -9,5 +11,6 @@ function CoffeeMachine:update(dt)
 end
 
 function CoffeeMachine:render()
-    love.graphics.draw(gFrames['CoffeeMachine'], self.x, self.y, 0, 0.5, 0.5)
+    love.graphics.draw(self.frame, self.x, self.y, 0, 
+        self.desired_width / self.frame:getWidth(), self.desired_height / self.frame:getHeight())
 end