+++ src/states/entity/Customer.lua
@@ -1,1 +1,0 @@
-Customer = class {__includes = BaseState}